Move validation code to appropriate model class
